==> script1_data_preparation_20240327.Rmd <==
---
title: "BHAMM Sex Differences Segmented Regression Study"
subtitle: "Script 1 - Data Preparation"
author: "Annalise LaPlume (adapted code from original author Rikki Lissaman)"
date of original script: "26-09-2023"
date of edited script: "`r format(Sys.time(), '%B %d, %Y')`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

# Overview
This R Markdown document contains code that imports BHAMM data, cleans it, and subsequently outputs files for analysis.

Inputs:

  * SPSS file of BHAMM master database (Last updated: Mar 2024) 
  * SPSS file of BHAMM session 2 database (Last updated: Oct 2023)

Outputs:

  * SPSS file containing behavioral data for the Segmented project

# Packages

Load packages.
```{r load}
library(haven) # version 2.5.0
library(readxl) # version 1.4.0
library(janitor) # version 2.1.0
library(tidyr) # version 1.2.0
library(dplyr) # version 1.0.8
library(psych) # version 2.3.3
```

# Set working directory to current folder
```{r wd}
setwd("C:/Users/annal/OneDrive/Documents/BHAMM/Segmented Regression/Sex Differences")
```

# Import and Combine Data

Import the BHAMM study master database and "clean" the variable names (i.e., make them all lower-case with spaces replaced by underscores).
```{r import data}
master <- read_sav("BHAMM_Master-All_V2024-01_Final.sav") %>% clean_names() # n = 480

```


Import the BHAMM session 2 database, "clean" the variable names, and filter out variables that are already in the master database (note: the master database is more comprehensive).
```{r import session 2}
session_two <- read_sav("BHAMM_S2_NOTES_V2023-10.sav") %>% 
  clean_names() %>%
  dplyr::select(-c("s2_date", "s2_tester", "s2_meno_group", 
            "s2_age", "s2_group")) # n = 261
```



Using participant ID, join together the two databases. Only those with session 2 data will be retained.
```{r merge}
bhamm <- left_join(session_two, master, by = "id") # n = 261
```

# Exclude participants based on background/demographic criteria

Filter the joined database by age, excluding older adults (>age 65).
```{r exclude older adults}
dplyr::count(subset(bhamm, s1_age > 65.5)) # n=2 excluded

bhamm <- bhamm %>% filter(s1_age < 65.5) 
```

Given our focus on sex and menopause, filter out participants whose current gender identity does not match their sex assigned at birth. Note: for sex, labels are "-1 [Male]" and "1 [Female]"; for gender, labels are "1 [Man]" and "2 [Woman]".
```{r exclude gender}

bhamm <- filter(bhamm, (sex == -1 & enrol_gender == 1) | (sex == -1 & is.na(enrol_gender)) | (sex == 1 & enrol_gender == 2) | (sex == 1 & is.na(enrol_gender))) # n = 1 excluded
```



Filter out participants whose menopause status is listed as perimenopausal ("2 [Perimenopause]" under `s2_meno_group`) or indeterminate ("4 [Indeterminate] under `s2_meno_group`). 
```{r exclude meno}
bhamm <- filter(bhamm, !s2_meno_group %in% c(2, 4)) # n = 47 excluded
```

Filter out participants currently using hormone replacement therapy ("2 [Yes, currently]" under `hrt_use`).
```{r exclude hrt}
bhamm <- filter(bhamm, !hrt_use %in% c(2)) # n = 10 excluded
```

Filter out participants with diabetes ("-1 [Yes]" under `med_con_diabetes`).
```{r exclude diabetes}
bhamm <- filter(bhamm, !med_con_diabetes %in% c(-1)) # n = 1 excluded
```

Filter out participants who are currently using hormonal birth control for a reason other than contraception ("1 [No]" under `current_hbc_med_use_contraception`). Note: those not using birth control are listed as NA so filtering in this way does not remove these individuals.
```{r exclude hbc}
bhamm <- filter(bhamm, !current_hbc_med_use_contraception %in% c(1)) 
# n = 4 excluded
```

Filter out participants who are currently using hormonal birth control and postmenopausal 

```{r exclude hbc postmeno}

table(bhamm$s2_meno_group, bhamm$current_hbc_med_use_contraception)

bhamm <- filter(bhamm, !(current_hbc_med_use_contraception %in% c(-1) &
                                s2_meno_group %in% c(3))) 
# n = 2 excluded
```

Filter out participants who have a BMI of 40 or above.
```{r exclude bmi}
bhamm <- filter(bhamm, bmi < 40) # n = 1 excluded
```


Filter out participants with neurological conditions ("-1 [Yes]" under `med_con_neurological`).
```{r exclude neurol}
bhamm <- filter(bhamm, !med_con_neurological %in% c(-1))  # n = 1 excluded

```

# Save dataset for CVLT analysis 
Output the full dataset as an SPSS file.
```{r output cvlt data}
bhamm_all <- bhamm

write_sav(bhamm_all, "BHAMM_Behav_Data_For_Analysis_cvlt_2024_03.sav")
```


## Descriptive statistics by menopause group

```{r meno describe}
# n per group
bhamm %>% 
  filter(sex == 1) %>%
  mutate(s2_meno_group = as_factor(s2_meno_group)) %>%
  count(s2_meno_group)

# Age by meno group
bhamm %>% 
  filter(sex == 1) %>%
  mutate(s2_meno_group = as_factor(s2_meno_group)) %>%
  group_by(s2_meno_group) %>%
  summarise(mean = round(mean(s2_age), 2),
            sd = round(sd(s2_age), 2),
            min = round(min(s2_age), 2),
            max = round(max(s2_age), 2))

# Education by meno group
bhamm %>% 
  filter(sex == 1) %>%
  mutate(s2_meno_group = as_factor(s2_meno_group)) %>%
  group_by(s2_meno_group) %>%
  summarise(mean = round(mean(edu), 2),
            sd = round(sd(edu), 2),
            min = round(min(edu), 2),
            max = round(max(edu), 2))

```


# Descriptives 

Descriptives for complete sample

## n 
```{r n all, include=TRUE, warning=FALSE, message=FALSE}
count(bhamm_all)
```


## Age
```{r age, include=TRUE, warning=FALSE, message=FALSE}
# Age 
describe(bhamm_all$s2_age)
```


## Ethnicity
```{r ethnicity, include=TRUE, warning=FALSE, message=FALSE}
# n per ethnicity
table(bhamm_all$ethnicity)
# 1=Aboriginal person, 2=Black, 3=Southeast Asian, 4=South Asian, 
# 5=West Asian, 6=Arab, 7=Chinese, 8=White, 9=Latin American, 
# 10=Filipino, 11=Korean, 12=Japanese, 13_White Southeast Asian, 
# 14=White Black, 15=White Ashkenazi, 16=White Latin American,
# 19=North African, 21=White + Aboriginal, 22 = White + Arab

# Calculate percentages for ethnicity data
# 1: indigenous
1/192*100 # 0.5%

# 2: black
6/192*100 # 3.1%

# 3: southeast asian + 10: filipino
(1+1)/192*100 # 1.0%

# 4: south asian 
7/192*100 # 3.6%

# 5: west asian + 6: arab
(1+4)/192*100 # 2.6%

# 8: white 
143/192*100 # 74.5%

# 9: latin american
9/192*100 # 4.7%

# east asian, 7: chinese + 11:korean + 12: japanese 
(11+0+1)/192*100 # 6.3%

# mixed, 14: white black + 16: white latin american 
# + 21:White Aboriginal, 22:White Arab
(1+3+1+1)/192*100 # 3.1%

# 19: north african
1/192*100 # 0.5

# total ethnicity data (rounded to 1 dp)
0.5 + 3.1 + 1 + 3.6 + 2.6 + 74.5 + 4.7 + 6.3 + 3.1 + 0.5

# total ethnicity data (rounded to whole numbers)
1 + 3 + 1 + 4 + 3 + 75 + 5 + 6 + 3 + 1

```


## Sex

Descriptive statistics per sex

```{r sex desc, include=TRUE, warning=FALSE, message=FALSE}
# Table 1

# n per sex
table(bhamm_all$sex)

# Age by sex
describeBy(bhamm_all$s2_age, bhamm_all$sex, mat = TRUE) 
# Check that groups do not differ on background variables (t-test)
t.test(bhamm_all$s2_age ~  bhamm_all$sex)

# Education by sex
describeBy(bhamm_all$edu, bhamm_all$sex, mat = TRUE) 
t.test(bhamm_all$edu ~  bhamm_all$sex)
```

## Menopause group

Descriptive statistics per meno group

```{r meno desc, include=TRUE, warning=FALSE, message=FALSE}

# Table 2
# n per menopause group
table(bhamm_all$s2_meno_group)

# Age by meno group
describeBy(bhamm_all$s2_age, bhamm_all$s2_meno_group, mat = TRUE) 
t.test(bhamm_all$s2_age ~  bhamm_all$s2_meno_group)

# Education by meno group
describeBy(bhamm_all$edu, bhamm_all$s2_meno_group, mat = TRUE) 
t.test(bhamm_all$edu ~  bhamm_all$s2_meno_group)
```


# Apply Additional Exclusions for Spatial Context Memory Task Performance


Filter out participants who did not provide at least one source correct response (across task versions). 
```{r exclude cs}

bhamm$id[which(bhamm$cs_rate_easy == 0)] # ids 3046 4015 5009 5069 7023
bhamm$id[which(bhamm$cs_rate_hard == 0)] # id 3046 4015 5009 5069 7023

bhamm <- filter(bhamm, cs_rate_easy != 0 & cs_rate_hard != 0) # n = 11 

```

Filter out participants who performed below chance for correct rejections (1/6 = 0.16).
```{r exclude cr}
bhamm <- filter(bhamm, cr_rate_easy >= 0.16 & cr_rate_hard >= 0.16) # n = 1
```


Change task variable types to numeric.
```{r str num}
bhamm <- bhamm %>%
  mutate_at(vars(ends_with("easy")), as.numeric) %>%
  mutate_at(vars(ends_with("hard")), as.numeric) %>%
  mutate(s2_age = as.numeric(s2_age))
```

Run models regressing age on correct source (a.k.a., spatial context retrieval accuracy), recognition, and misses.
```{r regression}
lm_cs_easy <- lm(cs_rate_easy ~ s2_age, data = bhamm)
lm_cs_hard <- lm(cs_rate_hard ~ s2_age, data = bhamm)
lm_recog_easy <- lm(recog_rate_easy ~ s2_age, data = bhamm)
lm_recog_hard <- lm(recog_rate_hard ~ s2_age, data = bhamm)
lm_miss_easy <- lm(misses_rate_easy ~ s2_age, data = bhamm)
lm_miss_hard <- lm(misses_rate_hard ~ s2_age, data = bhamm)
```

Calculate Cook's D for each regression model and store.
```{r cooks d}
cooks_d <- data.frame(id = bhamm$id, 
                      cs_easy_d = cooks.distance(lm_cs_easy),
                      cs_hard_d = cooks.distance(lm_cs_hard),
                      recog_easy_d = cooks.distance(lm_recog_easy),
                      recog_hard_d = cooks.distance(lm_recog_hard),
                      miss_easy_d = cooks.distance(lm_miss_easy),
                      miss_hard_d = cooks.distance(lm_miss_hard)) # n = 93

head(cooks_d)
```

For each measure (correct source, recognition, misses), determine whether a given participant's Cook D value is more than 3 standard deviations higher than the mean. After calculating these values, sum across the outlier columns. 
```{r outliers}
cooks_d <- cooks_d %>%
  mutate(cs_easy_outlier = case_when(cs_easy_d >= (mean(cs_easy_d) + 3*sd(cs_easy_d)) ~ 1,
                                     cs_easy_d < (mean(cs_easy_d) + 3*sd(cs_easy_d)) ~ 0),
         cs_hard_outlier = case_when(cs_hard_d >= (mean(cs_hard_d) + 3*sd(cs_hard_d)) ~ 1,
                                     cs_hard_d < (mean(cs_hard_d) + 3*sd(cs_hard_d)) ~ 0),
         recog_easy_outlier = case_when(recog_easy_d >= (mean(recog_easy_d) + 3*sd(recog_easy_d)) ~ 1,
                                     recog_easy_d < (mean(recog_easy_d) + 3*sd(recog_easy_d)) ~ 0),
         recog_hard_outlier = case_when(recog_hard_d >= (mean(recog_hard_d) + 3*sd(recog_hard_d)) ~ 1,
                                     recog_hard_d < (mean(recog_hard_d) + 3*sd(recog_hard_d)) ~ 0),
         miss_easy_outlier = case_when(miss_easy_d >= (mean(miss_easy_d) + 3*sd(miss_easy_d)) ~ 1,
                                     miss_easy_d < (mean(miss_easy_d) + 3*sd(miss_easy_d)) ~ 0),
         miss_hard_outlier = case_when(miss_hard_d >= (mean(miss_hard_d) + 3*sd(miss_hard_d)) ~ 1,
                                     miss_hard_d < (mean(miss_hard_d) + 3*sd(miss_hard_d)) ~ 0)) %>%
  mutate(total_outliers = rowSums(across(ends_with("outlier")))) 
```

Count participans that are outliers on a single measure
```{r count outliers}
# Check if individual measures have Cook's D >= 1 (definitely influential)
cooks_d %>% count(cs_easy_d >= 1)
cooks_d %>% count(cs_hard_d >= 1)
cooks_d %>% count(recog_easy_d >= 1)
cooks_d %>% count(recog_hard_d >= 1)
# no measure with Cook's D>1

# Check if individual measures have Cook's D >= 0.5 (possibly influential)
cooks_d %>% count(cs_easy_d >= 0.5)
cooks_d %>% count(cs_hard_d >= 0.5)
cooks_d %>% count(recog_easy_d >= 0.5)
cooks_d %>% count(recog_hard_d >= 0.5)
# no measure with Cook's D>0.5

# Check if individual measures have Cook's D > 3 SD above mean
cooks_d %>% count(cs_easy_outlier >= 1)
cooks_d %>% count(cs_hard_outlier >= 1)
cooks_d %>% count(recog_easy_outlier >= 1)
cooks_d %>% count(recog_hard_outlier >= 1)

# Find IDs of outliers per response measure
cooks_d$id[which(cooks_d$cs_easy_outlier >= 1)] # 1019 1036 1062 1074 2073 5001
cooks_d$id[which(cooks_d$cs_hard_outlier >= 1)] # 1049 1079 1089 2073 4045
cooks_d$id[which(cooks_d$recog_easy_outlier >= 1)] # 1074 4031 5001 5055 7013 7024
cooks_d$id[which(cooks_d$recog_hard_outlier >= 1)] # 1074 6013 7013 7036
```

Determine whether there are any participants deemed to be outliers on 3 or more measures.
```{r multiple outliers}
cooks_d %>% count(total_outliers >= 3)
```

Merge Cook's D values with dataset

```{r merge outliers}
# Merge data files
bhamm = merge(
  bhamm[,c("id", "s2_age", "s2_meno_group","ethnicity","edu","sex",
           "cs_rate_easy", "cs_rate_hard", 
           "recog_rate_easy", "recog_rate_hard", 
           "rt_cs_easy", "rt_cs_hard", "rt_recog_easy", "rt_recog_hard",
           "cvlt_lfr", "cvlt_lcr", "cvlt_rg", "enrol_tongue")],
  cooks_d[,c("id", "cs_easy_d","cs_hard_d" ,"recog_easy_d","recog_hard_d",
             "cs_easy_outlier","cs_hard_outlier",
             "recog_easy_outlier","recog_hard_outlier",
             "total_outliers")])
```

Remove participants with Cook's D > 3 SD above mean on 3 or more measures

```{r exclude outliers}
bhamm <- filter(bhamm, total_outliers != 3) # n = 2

# Remove participants on individual measures
# bhamm_alt <- filter(bhamm_outliers, cs_easy_outlier != 1) # n = 1
# bhamm_alt <- filter(bhamm_alt, cs_hard_outlier != 1) # n = 1
# bhamm_alt <- filter(bhamm_alt, recog_easy_outlier != 1) # n = 1
# bhamm_alt <- filter(bhamm_alt, recog_hard_outlier != 1) # n = 1
```

# Examine Basic Demographic Data for the Spatial Context Memory Task Sample

## n
```{r n}
dplyr::count(bhamm) # n = 178
```

## Age statistics (min, max, M, SD).
```{r age desc}
# age for whole sample
describe(bhamm$s2_age)
```

## Sex
```{r sex n}
bhamm %>% 
  mutate(sex = as_factor(sex)) %>%
  count(sex)
```

## Education
Calculate mean years of education.
```{r edu n}
bhamm %>%
  mutate(sex = as_factor(sex)) %>%
  group_by(sex) %>%
  summarise(mean = round(mean(edu), 2))
```

## Ethnicity
Count the different ethnicities included.
```{r ethnicity n}
bhamm %>% 
  mutate(sex = as_factor(sex)) %>%
  mutate(ethnicity = as_factor(ethnicity)) %>%
  group_by(sex) %>%
  count(ethnicity)
```

## Descriptive statistics by sex

```{r sex describe}
# Age by sex
bhamm %>% 
  mutate(sex = as_factor(sex)) %>%
  group_by(sex) %>%
  summarise(mean = round(mean(s2_age), 2),
            sd = round(sd(s2_age), 2),
            min = round(min(s2_age), 2),
            max = round(max(s2_age), 2))

# Education by sex
bhamm %>% 
  mutate(sex = as_factor(sex)) %>%
  group_by(sex) %>%
  summarise(mean = round(mean(edu), 2),
            sd = round(sd(edu), 2),
            min = round(min(edu), 2),
            max = round(max(edu), 2))

```





# Output Data For Analysis

Output the full dataset as an SPSS file.
```{r output all data}
write_sav(bhamm, "BHAMM_Behav_Data_For_Analysis_2024_03.sav")
```
