==> script3_verbal_memory_analyses_20240531.Rmd <==
---
title: "BHAMM Sex Differences Segmented Regression Study"
subtitle: "Script 3 - Analyses Verbal Memory Task"
author: "Annalise LaPlume"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output:
  html_document:
    highlight: textmate
    theme: paper
    toc: yes
    toc_depth: 3
    toc_float: yes
  pdf_document:
    toc: yes
editor_options:
  chunk_output_type: console
---
# Data setup

## Load data file and packages
```{r setup, include=TRUE, warning=FALSE, message=FALSE}

# Install and load packages
if(!require("haven"))install.packages("haven") 
if(!require("psych"))install.packages("psych") 
if(!require("dplyr"))install.packages("dplyr") 
if(!require("segmented"))install.packages("segmented") 
if(!require("ggplot2"))install.packages("ggplot2") 
if(!require("janitor"))install.packages("janitor") 
if(!require("forcats"))install.packages("forcats") 

library(haven) # load spss files
library(psych) # basic statistics functions
library(dplyr) # basic statistics functions
library(segmented) # segmented regression
library(ggplot2) # plot results
library(janitor) # examining and cleaning data
library(forcats) # working with categorical variables


# Set working directory to current folder
# Edit to specify where datafile is located
setwd("C:/Users/annal/OneDrive/Documents/BHAMM/Segmented Regression/Sex Differences")

# Import data
bhamm_cvlt <- read_sav("BHAMM_Behav_Data_For_Analysis_cvlt_2024_03.sav")

# view data sample
head(bhamm_cvlt)

names(bhamm_cvlt)
```

## Language variables 

```{r language, include=TRUE, warning=FALSE, message=FALSE}

# Mother tongue
str(bhamm_cvlt$enrol_tongue)
bhamm_cvlt$enrol_tongue <- as.factor(bhamm_cvlt$enrol_tongue)
table(bhamm_cvlt$enrol_tongue)

# '1_English', '2_French', '3_English & French', '4_English & Italian', 
# '5_Arabic', '6_Chinese', '7_Créole', '8_Farsi', '9_Greek', '10_Portuguese', 
#'11_Romanian', '12_Russian', '13_Spanish', '14_Tamil', '15_Italian', 
#'16_Bengali', '17_Bulgarian', '18_Telugu'
```

## Dependent variables

```{r dvs, include=TRUE, warning=FALSE, message=FALSE}
# cvlt_lfr: CVLT-II Long Free Recall Number                                   
# cvlt_lcr: CVLT-II Long Cued Recall Number                                        
# cvlt_rg: CVLT-II Recognition Number                                          

# The California Verbal Learning Task (CVLT)
# DVs: cvlt_lfr, cvlt_lcr, cvlt_rg
# Long Free Recall, Long Cued Recall, Recognition
# Predict that Free Recall would show the largest deficit

# Check variable classification
str(bhamm_cvlt$cvlt_lfr)
str(bhamm_cvlt$cvlt_lcr)
str(bhamm_cvlt$cvlt_rg)

# Describe variables
describe(bhamm_cvlt$cvlt_lfr)
describe(bhamm_cvlt$cvlt_lcr)
describe(bhamm_cvlt$cvlt_rg)

# Convert to percentages
bhamm_cvlt$cvlt_lfr_percent <- (bhamm_cvlt$cvlt_lfr/16)*100
bhamm_cvlt$cvlt_lcr_percent <- (bhamm_cvlt$cvlt_lcr/16)*100
bhamm_cvlt$cvlt_rg_percent <- (bhamm_cvlt$cvlt_rg/16)*100

# Describe percentages
describe(bhamm_cvlt$cvlt_lfr_percent)
describe(bhamm_cvlt$cvlt_lcr_percent)
describe(bhamm_cvlt$cvlt_rg_percent)

```

## Independent variables
```{r ivs, include=TRUE, warning=FALSE, message=FALSE}
# Convert sex and menopause status into factors
bhamm_cvlt$sex <- as.factor(bhamm_cvlt$sex)
bhamm_cvlt$s2_meno_group <- as.factor(bhamm_cvlt$s2_meno_group)

levels(bhamm_cvlt$s2_meno_group)
levels(bhamm_cvlt$s2_meno_group) <- c("Premenopause", "Postmenopause")
levels(bhamm_cvlt$s2_meno_group)

# Count n per menopause group
table(bhamm_cvlt$s2_meno_group)

# Sex
str(bhamm_cvlt$sex)
bhamm_cvlt$sex <- as.factor(bhamm_cvlt$sex)
levels(bhamm_cvlt$sex) <- c("Male", "Female")
table(bhamm_cvlt$sex)

# Add NA as a level to meno group factor 
bhamm_cvlt$s2_meno_group_na <- addNA(bhamm_cvlt$s2_meno_group)

# Sex by menopause status
table(bhamm_cvlt$sex, bhamm_cvlt$s2_meno_group_na)

# Age
str(bhamm_cvlt$s2_age)
describe(bhamm_cvlt$s2_age)

# Convert age into factor (grouped by decade)
bhamm_cvlt$age_decade <- cut(
  bhamm_cvlt$s2_age,
  breaks = c(0, 30, 40, 50, 60, 70),
   labels = c("0-29", "30-39", "40-49", "50-59", "60-65"),
  right  = FALSE)
```

## Sex differences factor
```{r sex diffs f, include=TRUE, warning=FALSE, message=FALSE}
# Convert sex and menopause status into a single factor
bhamm_cvlt$sex_meno_group <- fct_cross(bhamm_cvlt$sex, 
                                   bhamm_cvlt$s2_meno_group_na, 
                                   keep_empty = TRUE)

str(bhamm_cvlt$sex_meno_group)
levels(bhamm_cvlt$sex_meno_group)

# Collapse into single factor for males
bhamm_cvlt$sex_meno <- fct_collapse(bhamm_cvlt$sex_meno_group, 
                                      Male = c("Male:Premenopause",
                                                "Male:Postmenopause",
                                                "Male:NA"))
# Exclude NA females
bhamm_cvlt <- subset(bhamm_cvlt, sex_meno != "Female:NA")

table(bhamm_cvlt$sex_meno)
```


## Exclude participants whose first language is not English or French
```{r exclude land, include=TRUE, warning=FALSE, message=FALSE}
table(bhamm_cvlt$enrol_tongue)
bhamm_cvlt$l1 <- as.numeric(bhamm_cvlt$enrol_tongue)

bhamm_lang <- subset(bhamm_cvlt, l1<3)

table(bhamm_lang$enrol_tongue)

count(bhamm_cvlt) - count (bhamm_lang) # 27 excluded
```

## Separate datasets for males and females

```{r males females, include=TRUE, warning=FALSE, message=FALSE}
bhamm_lang_m <- subset(bhamm_lang, sex == "Male")
dplyr::count(bhamm_lang_m)

bhamm_lang_f <- subset(bhamm_lang, sex == "Female")
dplyr::count(bhamm_lang_f)
```


# Scatterplots

```{r cvlt scatterplots, include=TRUE, warning=FALSE, message=FALSE}
# Plot dependent variables by age, categorized by menopause group

# Long free recall

# save plot in jpeg format in current directory
jpeg(file="cvlt_lfr_scatterplot_bhamm.jpeg", width = 4, height = 4, 
     units = 'in', res = 300)

ggplot(data=bhamm_lang,                                         
       aes(x=s2_age, y = cvlt_lfr_percent,                                       
           colour=sex_meno,                                        
           na.rm= TRUE)) +                                              
  geom_point(size = 2, alpha = .6,                                     
             position = position_jitter(height = 0.1)) +  
   ylim (0,100) +
  labs(x="Age", y="Accuracy (% correct)",
       colour="",                   
       title="Verbal Memory Task \n Long Free Recall") +                
      theme_classic() +                                            
     theme(axis.text=element_text(size=14),                          
          axis.title=element_text(size=14),                        
           plot.title=element_text(size=14, face="bold", hjust = 0.5),
          legend.text=element_text(size=10),
          legend.position="none") +
          geom_smooth(method = "lm")+
  scale_colour_manual(values=c("cornflowerblue", "brown1", "darkorange1"))

# call this function to save the file 
dev.off()


# Long cued recall

# save plot in jpeg format in current directory
jpeg(file="cvlt_lcr_scatterplot_bhamm.jpeg", width = 4, height = 4, 
     units = 'in', res = 300)

ggplot(data=bhamm_lang,                                         
       aes(x=s2_age, y = cvlt_lcr_percent,                                       
           colour=sex_meno,                                        
           na.rm= TRUE)) +                                              
  geom_point(size = 2, alpha = .6,                                     
             position = position_jitter(height = 0.1)) +  
    ylim (0,100) +
  labs(x="Age", y="Accuracy (% correct)",
       colour="",                   
       title="Verbal Memory Task \n Long Cued Recall") +                
      theme_classic() +                                            
    theme(axis.text=element_text(size=14),                          
          axis.title=element_text(size=14),                        
           plot.title=element_text(size=14, face="bold", hjust = 0.5),
          legend.text=element_text(size=10),
          legend.position="none") +
          geom_smooth(method = "lm")+
  scale_colour_manual(values=c("cornflowerblue", "brown1", "darkorange1"))

# call this function to save the file 
dev.off()

# Recognition

# save plot in jpeg format in current directory
jpeg(file="cvlt_rg_scatterplot_bhamm.jpeg", width = 4, height = 4, 
     units = 'in', res = 300)

ggplot(data=bhamm_lang,                                         
       aes(x=s2_age, y = cvlt_rg_percent,                                       
           colour=sex_meno,                                        
           na.rm= TRUE)) +                                              
  geom_point(size = 2, alpha = .6,
             position = position_jitter(height = 0.1)) +  
    ylim (0,100) +
  labs(x="Age", y="Accuracy (% correct)",
       colour="",                   
       title="Verbal Memory Task \n Recognition") +                
      theme_classic() +                                            
     theme(axis.text=element_text(size=14),                          
          axis.title=element_text(size=14),                        
           plot.title=element_text(size=14, face="bold", hjust = 0.5),
          legend.text=element_text(size=10),
          legend.position="none") +
          geom_smooth(method = "lm")+
  scale_colour_manual(values=c("cornflowerblue", "brown1", "darkorange1"))

# call this function to save the file 
dev.off()

```


# Segmented regression

## Long free recall - Males

```{r segmented cvlt_lfr m, include=TRUE, warning=FALSE, message=FALSE}
# linear model
cvlt_lfr_m_lm <- lm(cvlt_lfr_percent ~ s2_age + edu, data=bhamm_lang_m)
summary(cvlt_lfr_m_lm)

# test for breakpoint
davies.test(cvlt_lfr_m_lm, ~ s2_age)

# segmented model with one breakpoint
cvlt_lfr_m_seg_1 <- segmented(cvlt_lfr_m_lm, ~ s2_age,
                           psi=50,
                           data=bhamm_lang_m)

summary(cvlt_lfr_m_seg_1)

# Model comparison: linear vs segmented
anova(cvlt_lfr_m_lm, cvlt_lfr_m_seg_1)
AIC(cvlt_lfr_m_lm, cvlt_lfr_m_seg_1)

# slope
slope(cvlt_lfr_m_seg_1)

# confidence intervals
cvlt_lfr_m_seg_1$psi

# plot.segmented(cvlt_lfr_m_seg_1)
```

## Long free recall - Females
```{r segmented cvlt_lfr f, include=TRUE, warning=FALSE, message=FALSE}
# linear model
cvlt_lfr_f_lm <- lm(cvlt_lfr_percent ~ s2_age + edu, data=bhamm_lang_f)
summary(cvlt_lfr_f_lm)

# test for breakpoint
davies.test(cvlt_lfr_f_lm, ~ s2_age)

# segmented model with one breakpoint
cvlt_lfr_f_seg_1 <- segmented(cvlt_lfr_f_lm, ~ s2_age,
                           psi=50,
                           data=bhamm_lang_f)

summary(cvlt_lfr_f_seg_1)

# Model comparison: linear vs segmented
anova(cvlt_lfr_f_lm, cvlt_lfr_f_seg_1)
AIC(cvlt_lfr_f_lm, cvlt_lfr_f_seg_1)

# slope
slope(cvlt_lfr_f_seg_1)

# confidence intervals
cvlt_lfr_f_seg_1$psi

# plot.segmented(cvlt_lfr_f_seg_1)
```

## Long cued recall - Males
```{r segmented cvlt_lcr m, include=TRUE, warning=FALSE, message=FALSE}
# linear model
cvlt_lcr_m_lm <- lm(cvlt_lcr_percent ~ s2_age + edu, data=bhamm_lang_m)
summary(cvlt_lcr_m_lm)

# test for breakpoint
davies.test(cvlt_lcr_m_lm, ~ s2_age)

# segmented model with one breakpoint
# os <-segmented(o, ~xg0+xg1, psi=list(xg0=.5,xg1=.5))
cvlt_lcr_m_seg_1 <- segmented(cvlt_lcr_m_lm, ~ s2_age,
                           psi=50,
                           data=bhamm_lang_m)

summary(cvlt_lcr_m_seg_1)

# Model comparison: linear vs segmented
anova(cvlt_lcr_m_lm, cvlt_lcr_m_seg_1)
AIC(cvlt_lcr_m_lm, cvlt_lcr_m_seg_1)

# slope
slope(cvlt_lcr_m_seg_1)

# confidence intervals
cvlt_lcr_m_seg_1$psi

# plot.segmented(cvlt_lcr_m_seg_1)
```

## Long cued recall - Females
```{r segmented cvlt_lcr f, include=TRUE, warning=FALSE, message=FALSE}
# linear model
cvlt_lcr_f_lm <- lm(cvlt_lcr_percent ~ s2_age + edu, data=bhamm_lang_f)
summary(cvlt_lcr_f_lm)

# test for breakpoint
davies.test(cvlt_lcr_f_lm, ~ s2_age)
davies.test(cvlt_lcr_f_lm, ~ s2_age)

# segmented model with one breakpoint
cvlt_lcr_f_seg_1 <- segmented(cvlt_lcr_f_lm, ~ s2_age,
                           psi=50,
                           data=bhamm_lang_f)

summary(cvlt_lcr_f_seg_1)

# Model comparison: linear vs segmented
anova(cvlt_lcr_f_lm, cvlt_lcr_f_seg_1)
AIC(cvlt_lcr_f_lm, cvlt_lcr_f_seg_1)

# slope
slope(cvlt_lcr_f_seg_1)

# confidence intervals
cvlt_lcr_f_seg_1$psi

# plot.segmented(cvlt_lcr_f_seg_1)
```

## Recognition - Males

```{r segmented cvlt_rg m, include=TRUE, warning=FALSE, message=FALSE}
# linear model
cvlt_rg_m_lm <- lm(cvlt_rg_percent ~ s2_age + edu, data=bhamm_lang_m)
summary(cvlt_rg_m_lm)

# test for breakpoint
davies.test(cvlt_rg_m_lm, ~ s2_age)

# segmented model with one breakpoint
cvlt_rg_m_seg_1 <- segmented(cvlt_rg_m_lm, ~ s2_age,
                           psi=50,
                           data=bhamm_lang_m)

summary(cvlt_rg_m_seg_1)

# Model comparison: linear vs segmented
anova(cvlt_rg_m_lm, cvlt_rg_m_seg_1)
AIC(cvlt_rg_m_lm, cvlt_rg_m_seg_1)

# slope
slope(cvlt_rg_m_seg_1)

# confidence intervals
cvlt_rg_m_seg_1$psi

# plot.segmented(cvlt_rg_m_seg_1)
```


## Recognition - Females

```{r segmented cvlt_rg f, include=TRUE, warning=FALSE, message=FALSE}
# linear model
cvlt_rg_f_lm <- lm(cvlt_rg_percent ~ s2_age + edu, data=bhamm_lang_f)
summary(cvlt_rg_f_lm)

# test for breakpoint
davies.test(cvlt_rg_f_lm, ~ s2_age)

# segmented model with one breakpoint
cvlt_rg_f_seg_1 <- segmented(cvlt_rg_f_lm, ~ s2_age,
                           psi=50,
                           data=bhamm_lang_f)

summary(cvlt_rg_f_seg_1)

# Model comparison: linear vs segmented
anova(cvlt_rg_f_lm, cvlt_rg_f_seg_1)
AIC(cvlt_rg_f_lm, cvlt_rg_f_seg_1)

# slope
slope(cvlt_rg_f_seg_1)

# confidence intervals
cvlt_rg_f_seg_1$psi

# plot.segmented(cvlt_rg_f_seg_1)
```


# Best fitting models

Hierarchical model comparison
```{r model comparison, include=TRUE, warning=FALSE, message=FALSE}

# report model comparison p-values, F, and delta AIC in paper

# Long free recall
anova(cvlt_lfr_m_lm, cvlt_lfr_m_seg_1)
anova(cvlt_lfr_f_lm, cvlt_lfr_f_seg_1)

AIC(cvlt_lfr_m_seg_1) - AIC (cvlt_lfr_m_lm)
AIC(cvlt_lfr_f_seg_1) - AIC (cvlt_lfr_f_lm)

summary(cvlt_lfr_m_seg_1)$adj.r.squared - summary(cvlt_lfr_m_lm)$adj.r.squared
summary(cvlt_lfr_f_seg_1)$adj.r.squared - summary(cvlt_lfr_f_lm)$adj.r.squared

# Long cued recall
anova(cvlt_lcr_m_lm, cvlt_lcr_m_seg_1)
anova(cvlt_lcr_f_lm, cvlt_lcr_f_seg_1)

AIC(cvlt_lcr_m_seg_1) - AIC (cvlt_lcr_m_lm)
AIC(cvlt_lcr_f_seg_1) - AIC (cvlt_lcr_f_lm)

summary(cvlt_lcr_m_seg_1)$adj.r.squared - summary(cvlt_lcr_m_lm)$adj.r.squared
summary(cvlt_lcr_f_seg_1)$adj.r.squared - summary(cvlt_lcr_f_lm)$adj.r.squared

# Recognition
anova(cvlt_rg_m_lm, cvlt_rg_m_seg_1)
anova(cvlt_rg_f_lm, cvlt_rg_f_seg_1)

AIC(cvlt_rg_m_seg_1) - AIC (cvlt_rg_m_lm)
AIC(cvlt_rg_f_seg_1) - AIC (cvlt_rg_f_lm)

summary(cvlt_rg_m_seg_1)$adj.r.squared - summary(cvlt_rg_m_lm)$adj.r.squared
summary(cvlt_rg_f_seg_1)$adj.r.squared - summary(cvlt_rg_f_lm)$adj.r.squared

```

Estimates from best fitting models

```{r fitted estimates, include=TRUE, warning=FALSE, message=FALSE}
# report coefficient values for s2_age in paper (estimate, SE, t, p)

# Long free recall
summary(cvlt_lfr_m_lm)
summary(cvlt_lfr_f_lm)

# Long cued recall
summary(cvlt_lcr_m_lm)
summary(cvlt_lcr_f_lm)

# Recognition
summary(cvlt_rg_m_lm)
summary(cvlt_rg_f_lm)

```

Summary: Best fitting model for all measures in both sexes is linear
